Add `getConfigForLabel()`
In `cerebro`, you can assign tags (it calls them `labels`) to settings to group settings under a particular tag.

Although it could do this, its usage was very limited as it does not have a getter to work with them.

The new method allows you to get a set of settings by its tag.